requirements.txt
asgiref==3.5.2
async-generator==1.10
attrs==21.4.0
bcrypt==3.2.2
beautifulsoup4==4.11.1
bs4==0.0.1
cffi==1.15.0
charset-normalizer==2.1.0
cryptography==37.0.2
cycler==0.11.0
Django==4.0.5
django-bootstrap4==22.1
django-cors-headers==3.13.0
fonttools==4.33.3
h11==0.13.0
idna==3.3
jwt==1.3.1
kiwisolver==1.4.3
matplotlib==3.5.2
mysqlclient==2.1.1
numpy==1.23.0
outcome==1.2.0
packaging==21.3
pandas==1.4.3
Pillow==9.1.1
pycparser==2.21
PyMySQL==1.0.2
pyOpenSSL==22.0.0
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2022.1
requests==2.28.1
scipy==1.8.1
seaborn==0.11.2
selenium==4.3.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
trio==0.21.0
trio-websocket==0.9.2
tzdata==2022.1
urllib3==1.26.9
wincertstore==0.2
wsproto==1.1.0
